requirements.txt
google-cloud-monitoring
google-cloud-resource-manager
google-cloud-
google-cloud-artifact-registry